requirements.txt
Flask
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
passlib
marshmallow
python-dotenv
redis[hiredis]
psycopg2-binary
alembic
flask-migrate
bcrypt